Create FSM_StoreLoad.v
This is a modification of FSM_Wrapper in Lab 4.  Changes currently allow for the store instruction to be executed.
NOTES: 
Removed memory location requirement (L165) for testing purposes.  Will probably need to reinstate in final version
Added mem_mux module which writes to the addr_b of memory module and takes in the value from from Register A as the Address.  Will store the value of Register B in that address.
TODO:
Implement Load functionality